fix: should generate unique aside link ide when using different syntax
