+++ .changeset/cold-bulldogs-kiss.md
@@ -1,0 +1,5 @@
+---
+'@rspress/core': patch
+---
+
+release
