--- e2e/fixtures/custom-id/doc/guide.mdx
@@ -1,3 +1,0 @@
-# Guide
-
-## title 1 {#custom-id}
