+++ e2e/fixtures/toc-id/doc/index.mdx
@@ -1,0 +1,15 @@
+# Guide
+
+## title 1 {#custom-id}
+
+## title 2
+
+## title 2
+
+## `title` 2
+
+## Title 2
+
+## **title** 2
+
+## *title* 2
