+++ packages/core/package.json
@@ -46,7 +46,7 @@
     "@rsbuild/plugin-less": "~1.1.0",
     "@rsbuild/plugin-react": "~1.0.7",
     "@rsbuild/plugin-sass": "~1.1.1",
-    "@rspress/mdx-rs": "0.6.3",
+    "@rspress/mdx-rs": "0.6.4",
     "@rspress/plugin-auto-nav-sidebar": "workspace:*",
     "@rspress/plugin-container-syntax": "workspace:*",
     "@rspress/plugin-last-updated": "workspace:*",
